Add files via upload
DataScience-HarvardX-ChooseYourOwnProject is an in-depth data science project focusing on traffic volume forecasting on an American interstate using eXtreme Gradient Boosting (xgbTree) in R. Developed as part of a HarvardX course, this project embodies a thorough exploration of machine learning techniques, encompassing the iterative tuning of model hyperparameters, rigorous performance evaluation, and insightful feature importance analysis. The project provides a blend of theoretical knowledge and practical application, making it a valuable resource for students, researchers, and practitioners interested in applying advanced statistical methods to real-world problems.